Added the 1st and 2nd versions of the program.
There are still many quality of life things that I would like to do as well as some cleanup and comments. The files are main.py which is v1 of the program and main_v2.py which is v2 of the program. Then store.json stores the files name, path, size, and MD5 hash value.